requirements_scraper.txt
playwright==1.49.0
google-cloud-firestore>=2.16.0
pandas>=2.2.2
beautifulsoup4>=4.12.3
